+++ app/src/main/java/eu/faircode/email/FragmentCompose.java
@@ -1291,6 +1291,7 @@ public class FragmentCompose extends FragmentBase {
                         Bundle extras = new Bundle();
                         extras.putString("html", html);
                         extras.putBoolean("show", true);
+                        extras.putBoolean("refedit", true);
                         onAction(R.id.action_save, extras, "refedit");
                     }
 
@@ -1939,12 +1940,18 @@ public class FragmentCompose extends FragmentBase {
                 for (String group : groups)
                     map.put(group, main.addSubMenu(Menu.NONE, order, order++, group));
 
+                NumberFormat NF = NumberFormat.getNumberInstance();
                 for (EntityAnswer answer : answers) {
                     if (answer.favorite)
                         continue;
                     order++;
 
-                    String name = answer.name;
+                    SpannableStringBuilder name = new SpannableStringBuilder(answer.name);
+                    if (grouped && answer.applied > 0) {
+                        name.append(" (").append(NF.format(answer.applied)).append(")");
+                        name.setSpan(new RelativeSizeSpan(HtmlHelper.FONT_SMALL),
+                                answer.name.length() + 1, name.length(), 0);
+                    }
 
                     MenuItem item;
                     if (answer.group == null)
@@ -4935,6 +4942,8 @@ public class FragmentCompose extends FragmentBase {
                     }
 
                     if (save_drafts &&
+                            (data.draft.ui_encrypt == null ||
+                                    EntityMessage.ENCRYPT_NONE.equals(data.draft.ui_encrypt)) &&
                             (!"new".equals(action) ||
                                     answer > 0 ||
                                     !TextUtils.isEmpty(to) ||
@@ -5240,7 +5249,7 @@ public class FragmentCompose extends FragmentBase {
                                 };
                             }
 
-                            showDraft(draft, postShow);
+                            showDraft(draft, false, postShow);
                         }
 
                         tvDsn.setVisibility(
@@ -6019,8 +6028,9 @@ public class FragmentCompose extends FragmentBase {
 
             Bundle extras = args.getBundle("extras");
             boolean show = extras.getBoolean("show");
+            boolean refedit = extras.getBoolean("refedit");
             if (show)
-                showDraft(draft, null);
+                showDraft(draft, refedit, null);
 
             bottom_navigation.getMenu().findItem(R.id.action_undo).setVisible(draft.revision > 1);
             bottom_navigation.getMenu().findItem(R.id.action_redo).setVisible(draft.revision < draft.revisions);
@@ -6041,7 +6051,7 @@ public class FragmentCompose extends FragmentBase {
                 finish();
 
             } else if (action == R.id.action_undo || action == R.id.action_redo) {
-                showDraft(draft, null);
+                showDraft(draft, false, null);
 
             } else if (action == R.id.action_save) {
                 boolean autosave = extras.getBoolean("autosave");
@@ -6202,7 +6212,7 @@ public class FragmentCompose extends FragmentBase {
                 ref.first().before(div);
     }
 
-    private void showDraft(final EntityMessage draft, Runnable postShow) {
+    private void showDraft(final EntityMessage draft, boolean refedit, Runnable postShow) {
         Bundle args = new Bundle();
         args.putLong("id", draft.id);
         args.putBoolean("show_images", show_images);
@@ -6333,6 +6343,9 @@ public class FragmentCompose extends FragmentBase {
                 ibReferenceImages.setVisibility(ref_has_images && !show_images ? View.VISIBLE : View.GONE);
 
                 setBodyPadding();
+
+                if (refedit && write_below)
+                    etBody.setSelection(etBody.length());
 
                 if (state == State.LOADED)
                     return;