+++ app/src/main/java/eu/faircode/email/ServiceAuthenticator.java
@@ -136,15 +136,8 @@ public class ServiceAuthenticator extends Authenticator {
             Log.e(ex);
         }
 
-        if (expiration != null && expiration - keep_alive < new Date().getTime()) {
-            try {
-                refreshToken(true);
-            } catch (Throwable ex) {
-                Log.w(ex);
-            }
-
-            throw new IllegalStateException("Token refreshed");
-        }
+        if (expiration != null && expiration - keep_alive < new Date().getTime())
+            throw new IllegalStateException(Log.TOKEN_REFRESH_REQUIRED);
     }
 
     interface IAuthenticated {